pyproject.toml
[project]
name = "PersonaGraph"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "langgraph>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "python-multipart>=0.0.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.15.11",
]
[tool.hatch.build.targets.wheel]
packages = ["."]
